requirements.txt
astroid==2.4.1
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2
django-cors-headers==2.2.0
django-model-utils==4.0.0
djangorestframework==3.8.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
Pillow==7.1.1
protobuf==3.11.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.2
pytz==2019.1
requests==2.21.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.2
wrapt==1.12.1
